[FIX] change test

+++ cpp_module07/ex01/iter.hpp
@@ -6,7 +6,7 @@
 /*   By: jhurpy <jhurpy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/24 20:38:49 by jhurpy            #+#    #+#             */
-/*   Updated: 2024/07/25 00:25:18 by jhurpy           ###   ########.fr       */
+/*   Updated: 2024/07/31 23:04:37 by jhurpy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include <iostream>
 
 template <typename T>
-void iter(T *array, int length, void (*function)(T const &))
+void iter(T *array, size_t length, void (*function)(T const &))
 {
 	for (int i = 0; i < length; i++)
 		function(array[i]);
